Feat (quant/float): move float quant out of experimental phase
