+++ src/brevitas/quant/__init__.py
@@ -3,6 +3,10 @@
 
 from .binary import *
 from .fixed_point import *
+from .float import *
+from .float_quant_fnuz import *
+from .float_quant_ocp import *
+from .mx_quant_ocp import *
 from .none import *
 from .scaled_int import *
 from .shifted_scaled_int import *
