+++ src/brevitas_examples/common/generative/quantize.py
@@ -22,30 +22,30 @@ from brevitas.graph.quantize import layerwise_quantize
 from brevitas.inject.enum import RestrictValueType
 from brevitas.inject.enum import StatsOp
 from brevitas.quant.base import ParameterFromRuntimeZeroPoint
-from brevitas.quant.experimental.float import Fp8e4m3Act
-from brevitas.quant.experimental.float import Fp8e4m3ActPerTensorFloat
-from brevitas.quant.experimental.float import Fp8e4m3WeightPerChannelFloat
-from brevitas.quant.experimental.float import Fp8e4m3WeightPerTensorFloat
-from brevitas.quant.experimental.float_quant_fnuz import Fp8e4m3FNUZActPerTensorFloat
-from brevitas.quant.experimental.float_quant_fnuz import Fp8e4m3FNUZWeightPerChannelFloat
-from brevitas.quant.experimental.float_quant_fnuz import Fp8e4m3FNUZWeightPerTensorFloat
-from brevitas.quant.experimental.float_quant_ocp import Fp8e4m3OCPActPerTensorFloat
-from brevitas.quant.experimental.float_quant_ocp import Fp8e4m3OCPWeightPerChannelFloat
-from brevitas.quant.experimental.float_quant_ocp import Fp8e4m3OCPWeightPerTensorFloat
-from brevitas.quant.experimental.mx_quant_ocp import MXFloat8e4m3Act
-from brevitas.quant.experimental.mx_quant_ocp import MXFloat8e4m3Weight
-from brevitas.quant.experimental.mx_quant_ocp import MXFloat8e4m3WeightMSE
-from brevitas.quant.experimental.mx_quant_ocp import MXInt8Act
-from brevitas.quant.experimental.mx_quant_ocp import MXInt8Weight
-from brevitas.quant.experimental.mx_quant_ocp import MXInt8WeightMSE
-from brevitas.quant.experimental.mx_quant_ocp import ShiftedMXUInt8Weight
-from brevitas.quant.experimental.mx_quant_ocp import ShiftedMXUInt8WeightMSE
 from brevitas.quant.fixed_point import Int8ActPerTensorFixedPoint
 from brevitas.quant.fixed_point import Int8ActPerTensorFixedPointMSE
 from brevitas.quant.fixed_point import Int8WeightPerChannelFixedPoint
 from brevitas.quant.fixed_point import Int8WeightPerChannelFixedPointMSE
 from brevitas.quant.fixed_point import Int8WeightPerTensorFixedPoint
 from brevitas.quant.fixed_point import Int8WeightPerTensorFixedPointMSE
+from brevitas.quant.float import Fp8e4m3Act
+from brevitas.quant.float import Fp8e4m3ActPerTensorFloat
+from brevitas.quant.float import Fp8e4m3WeightPerChannelFloat
+from brevitas.quant.float import Fp8e4m3WeightPerTensorFloat
+from brevitas.quant.float_quant_fnuz import Fp8e4m3FNUZActPerTensorFloat
+from brevitas.quant.float_quant_fnuz import Fp8e4m3FNUZWeightPerChannelFloat
+from brevitas.quant.float_quant_fnuz import Fp8e4m3FNUZWeightPerTensorFloat
+from brevitas.quant.float_quant_ocp import Fp8e4m3OCPActPerTensorFloat
+from brevitas.quant.float_quant_ocp import Fp8e4m3OCPWeightPerChannelFloat
+from brevitas.quant.float_quant_ocp import Fp8e4m3OCPWeightPerTensorFloat
+from brevitas.quant.mx_quant_ocp import MXFloat8e4m3Act
+from brevitas.quant.mx_quant_ocp import MXFloat8e4m3Weight
+from brevitas.quant.mx_quant_ocp import MXFloat8e4m3WeightMSE
+from brevitas.quant.mx_quant_ocp import MXInt8Act
+from brevitas.quant.mx_quant_ocp import MXInt8Weight
+from brevitas.quant.mx_quant_ocp import MXInt8WeightMSE
+from brevitas.quant.mx_quant_ocp import ShiftedMXUInt8Weight
+from brevitas.quant.mx_quant_ocp import ShiftedMXUInt8WeightMSE
 from brevitas.quant.scaled_int import Int8ActPerTensorFloat
 from brevitas.quant.scaled_int import Int8ActPerTensorFloatMSE
 from brevitas.quant.scaled_int import Int8WeightPerChannelFloat
