+++ src/brevitas_examples/imagenet_classification/ptq/ptq_common.py
@@ -24,32 +24,32 @@ from brevitas.graph.quantize import quantize
 from brevitas.graph.target.flexml import quantize_flexml
 from brevitas.inject import value
 import brevitas.nn as qnn
-from brevitas.quant.experimental.float import Fp8e4m3ActPerTensorFloat
-from brevitas.quant.experimental.float import Fp8e4m3ActPerTensorFloatMSE
-from brevitas.quant.experimental.float import Fp8e4m3WeightPerChannelFloat
-from brevitas.quant.experimental.float import Fp8e4m3WeightPerChannelFloatMSE
-from brevitas.quant.experimental.float import Fp8e4m3WeightPerTensorFloat
-from brevitas.quant.experimental.float import Fp8e4m3WeightPerTensorFloatMSE
-from brevitas.quant.experimental.float_quant_ocp import Fp8e4m3OCPActPerTensorFloat
-from brevitas.quant.experimental.float_quant_ocp import Fp8e4m3OCPActPerTensorFloatMSE
-from brevitas.quant.experimental.float_quant_ocp import Fp8e4m3OCPWeightPerChannelFloat
-from brevitas.quant.experimental.float_quant_ocp import Fp8e4m3OCPWeightPerChannelFloatMSE
-from brevitas.quant.experimental.float_quant_ocp import Fp8e4m3OCPWeightPerTensorFloat
-from brevitas.quant.experimental.float_quant_ocp import Fp8e4m3OCPWeightPerTensorFloatMSE
-from brevitas.quant.experimental.mx_quant_ocp import MXFloat8e4m3Act
-from brevitas.quant.experimental.mx_quant_ocp import MXFloat8e4m3Weight
-from brevitas.quant.experimental.mx_quant_ocp import MXFloat8e4m3WeightMSE
-from brevitas.quant.experimental.mx_quant_ocp import MXInt8Act
-from brevitas.quant.experimental.mx_quant_ocp import MXInt8Weight
-from brevitas.quant.experimental.mx_quant_ocp import MXInt8WeightMSE
-from brevitas.quant.experimental.mx_quant_ocp import ShiftedMXUInt8Weight
-from brevitas.quant.experimental.mx_quant_ocp import ShiftedMXUInt8WeightMSE
 from brevitas.quant.fixed_point import Int8ActPerTensorFixedPoint
 from brevitas.quant.fixed_point import Int8ActPerTensorFixedPointMSE
 from brevitas.quant.fixed_point import Int8WeightPerChannelFixedPoint
 from brevitas.quant.fixed_point import Int8WeightPerChannelFixedPointMSE
 from brevitas.quant.fixed_point import Int8WeightPerTensorFixedPoint
 from brevitas.quant.fixed_point import Int8WeightPerTensorFixedPointMSE
+from brevitas.quant.float import Fp8e4m3ActPerTensorFloat
+from brevitas.quant.float import Fp8e4m3ActPerTensorFloatMSE
+from brevitas.quant.float import Fp8e4m3WeightPerChannelFloat
+from brevitas.quant.float import Fp8e4m3WeightPerChannelFloatMSE
+from brevitas.quant.float import Fp8e4m3WeightPerTensorFloat
+from brevitas.quant.float import Fp8e4m3WeightPerTensorFloatMSE
+from brevitas.quant.float_quant_ocp import Fp8e4m3OCPActPerTensorFloat
+from brevitas.quant.float_quant_ocp import Fp8e4m3OCPActPerTensorFloatMSE
+from brevitas.quant.float_quant_ocp import Fp8e4m3OCPWeightPerChannelFloat
+from brevitas.quant.float_quant_ocp import Fp8e4m3OCPWeightPerChannelFloatMSE
+from brevitas.quant.float_quant_ocp import Fp8e4m3OCPWeightPerTensorFloat
+from brevitas.quant.float_quant_ocp import Fp8e4m3OCPWeightPerTensorFloatMSE
+from brevitas.quant.mx_quant_ocp import MXFloat8e4m3Act
+from brevitas.quant.mx_quant_ocp import MXFloat8e4m3Weight
+from brevitas.quant.mx_quant_ocp import MXFloat8e4m3WeightMSE
+from brevitas.quant.mx_quant_ocp import MXInt8Act
+from brevitas.quant.mx_quant_ocp import MXInt8Weight
+from brevitas.quant.mx_quant_ocp import MXInt8WeightMSE
+from brevitas.quant.mx_quant_ocp import ShiftedMXUInt8Weight
+from brevitas.quant.mx_quant_ocp import ShiftedMXUInt8WeightMSE
 from brevitas.quant.scaled_int import Int8ActPerTensorFloat
 from brevitas.quant.scaled_int import Int8ActPerTensorFloatMSE
 from brevitas.quant.scaled_int import Int8WeightPerChannelFloat
